+++ red_ttp/certutil_webrequest.py
@@ -4,22 +4,24 @@
 # Description: Uses certutil.exe to download a file.
 
 import common
+import platform
 
 MY_DLL = common.get_path("bin", "mydll.dll")
 
 
 @common.dependencies(MY_DLL)
 def main():
-    # http server will terminate on main thread exit
-    # if daemon is True
-    server, ip, port = common.serve_web()
 
+    if platform.system() != 'Windows':
+        common.log("This script only runs on Windows.")
+        return common.UNSUPPORTED_RTA
+
+    server, ip, port = common.serve_web()
     uri = "bin/mydll.dll"
     target_file = "mydll.dll"
     common.clear_web_cache()
     url = "http://{ip}:{port}/{uri}".format(ip=ip, port=port, uri=uri)
     common.execute(["certutil.exe", "-urlcache", "-split", "-f", url, target_file])
-
     server.shutdown()
     common.remove_file(target_file)
